requirements.txt
Flask
openpyxl
opencv-python
pillow
pytesseract
requests